rename dataframes module to lower case
